+++ apps/astro/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @keystatic/test-astro
 
+## 0.0.53
+
+### Patch Changes
+
+- Updated dependencies [d5e079a]
+  - @keystatic/core@0.0.89
+
+## 0.0.52
+
+### Patch Changes
+
+- Updated dependencies [f1c2585]
+  - @keystatic/core@0.0.88
+
 ## 0.0.51
 
 ### Patch Changes