+++ apps/localization/CHANGELOG.md
@@ -1,5 +1,21 @@
 # localization
 
+## 0.1.58
+
+### Patch Changes
+
+- Updated dependencies [d5e079a]
+  - @keystatic/core@0.0.89
+
+## 0.1.57
+
+### Patch Changes
+
+- Updated dependencies [f1c2585]
+- Updated dependencies [f1c2585]
+  - @keystatic/core@0.0.88
+  - @keystatic/next@0.0.7
+
 ## 0.1.56
 
 ### Patch Changes