+++ packages/keystatic/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @keystatic/core
 
+## 0.0.89
+
+### Patch Changes
+
+- d5e079a: Remove circular dependencies causing broken build
+
+## 0.0.88
+
+### Patch Changes
+
+- f1c2585: Add `@keystatic/core/api/utils`
+
 ## 0.0.87
 
 ### Patch Changes